fix(android): recover join via link
