+++ actor-sdk/sdk-core-android/android-sdk/src/main/java/im/actor/sdk/controllers/dialogs/BaseDialogFragment.java
@@ -33,8 +33,6 @@ public abstract class BaseDialogFragment extends DisplayListFragment<Dialog, Dia
 
     private View emptyDialogs;
 
-    private String joinGroupUrl;
-
     public BaseDialogFragment() {
         setRetainInstance(true);
     }
@@ -97,30 +95,6 @@ public abstract class BaseDialogFragment extends DisplayListFragment<Dialog, Dia
     @Override
     public void onResume() {
         super.onResume();
-        if (joinGroupUrl != null && !joinGroupUrl.isEmpty()) {
-            String[] urlSplit = null;
-            if (joinGroupUrl.contains("join")) {
-                urlSplit = joinGroupUrl.split("/join/");
-            } else if (joinGroupUrl.contains("token")) {
-                urlSplit = joinGroupUrl.split("token=");
-            }
-            if (urlSplit != null) {
-                joinGroupUrl = urlSplit[urlSplit.length - 1];
-                execute(messenger().joinGroupViaToken(joinGroupUrl), R.string.invite_link_title, new CommandCallback<Integer>() {
-                    @Override
-                    public void onResult(Integer res) {
-                        getActivity().startActivity(Intents.openGroupDialog(res, true, getActivity()));
-                        getActivity().finish();
-                        joinGroupUrl = "";
-                    }
-
-                    @Override
-                    public void onError(Exception e) {
-                        joinGroupUrl = "";
-                    }
-                });
-            }
-        }
         messenger().onDialogsOpen();
     }
